perl-ptest: copy xconfig data from perl source directory
The following perl ptests:

  - dist/ExtUtils-ParseXS/t/001-basic
  - dist/ExtUtils-ParseXS/t/002-more
  - dist/ExtUtils-ParseXS/t/003-usage
  - cpan/ExtUtils-Constant/t/Constant
  - cpan/ExtUtils-MakeMaker/t/02-xsdynamic

are erroring out with:
| /usr/lib/perl/ptest/perl_langinfo.h:8:10: fatal error:
xconfig.h: No such file or directory
|     8 | #include "xconfig.h"

xconfig.h contains references to the build host architecture and was
removed by commit 2e0f30c46802 ("perl: do not install files that
contain build host specific data")

However, it is still included from various other places including
these tests, and we are still depending on build host architecture
data by including the patches from perl-cross recipe, a dependency
to perl recipe.

xconfig.h was added back as a copy step in the commit 5b18d7a0c048
("update 5.36.1 -> 5.38.0") but was not added back in perl-ptest
include file.

Borrowed the logic from the above commit to make a copy of config.h
that is specific to the target architecture. The changes in this
commit fixes the test failures.

Signed-off-by: Rajendra Desai <rajendra.desai@emerson.com>